feat: board, post 타입 정리
